Issue-299: Moved BaseRepository and Result to core module.
