+++ app/src/main/java/com/kirchhoff/movies/mapper/discover/DiscoverMapper.kt
@@ -1,5 +1,6 @@
 package com.kirchhoff.movies.mapper.discover
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.data.ui.main.UITv
@@ -7,7 +8,6 @@ import com.kirchhoff.movies.mapper.BaseMapper
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.main.NetworkMovie
 import com.kirchhoff.movies.networkdata.main.NetworkTv
-import com.kirchhoff.movies.repository.Result
 
 class DiscoverMapper : BaseMapper(), IDiscoverMapper {
