+++ app/src/main/java/com/kirchhoff/movies/mapper/movie/MovieDetailsMapper.kt
@@ -1,5 +1,6 @@
 package com.kirchhoff.movies.mapper.movie
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.movie.UICountry
 import com.kirchhoff.movies.data.ui.details.movie.UIMovieDetails
@@ -12,7 +13,6 @@ import com.kirchhoff.movies.networkdata.core.NetworkObjectWithName
 import com.kirchhoff.movies.networkdata.details.movie.NetworkMovieDetails
 import com.kirchhoff.movies.networkdata.details.movie.NetworkTrailer
 import com.kirchhoff.movies.networkdata.details.movie.NetworkTrailersList
-import com.kirchhoff.movies.repository.Result
 
 class MovieDetailsMapper(private val coreMapper: ICoreMapper) : BaseMapper(), IMovieDetailsMapper {
     override fun createUIMovieDetails(movieDetailsResult: Result<NetworkMovieDetails>): Result<UIMovieDetails> =
