+++ app/src/main/java/com/kirchhoff/movies/mapper/review/ReviewListMapper.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.mapper.review
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.review.UIReview
 import com.kirchhoff.movies.mapper.BaseMapper
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.details.review.NetworkReview
-import com.kirchhoff.movies.repository.Result
 
 class ReviewListMapper : BaseMapper(), IReviewListMapper {
