+++ app/src/main/java/com/kirchhoff/movies/repository/discover/IDiscoverRepository.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.repository.discover
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.data.ui.main.UITv
-import com.kirchhoff.movies.repository.Result
 
 interface IDiscoverRepository {
     suspend fun fetchMovies(page: Int): Result<UIPaginated<UIMovie>>
