+++ app/src/main/java/com/kirchhoff/movies/repository/person/IPersonsRepository.kt
@@ -1,10 +1,10 @@
 package com.kirchhoff.movies.repository.person
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.person.UIPersonCredits
 import com.kirchhoff.movies.data.ui.details.person.UIPersonDetails
 import com.kirchhoff.movies.data.ui.main.UIPerson
-import com.kirchhoff.movies.repository.Result
 
 interface IPersonsRepository {
     suspend fun fetchPopularPersons(page: Int): Result<UIPaginated<UIPerson>>
