+++ app/src/main/java/com/kirchhoff/movies/repository/person/PersonsRepository.kt
@@ -1,5 +1,7 @@
 package com.kirchhoff.movies.repository.person
 
+import com.kirchhoff.movies.core.repository.BaseRepository
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.person.UIPersonCredits
 import com.kirchhoff.movies.data.ui.details.person.UIPersonDetails
@@ -7,8 +9,6 @@ import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.mapper.person.details.IPersonDetailsMapper
 import com.kirchhoff.movies.mapper.person.main.IPersonsMapper
 import com.kirchhoff.movies.network.services.PersonService
-import com.kirchhoff.movies.repository.BaseRepository
-import com.kirchhoff.movies.repository.Result
 
 class PersonsRepository(
     private val personService: PersonService,
