+++ app/src/main/java/com/kirchhoff/movies/repository/tv/TvRepository.kt
@@ -1,5 +1,7 @@
 package com.kirchhoff.movies.repository.tv
 
+import com.kirchhoff.movies.core.repository.BaseRepository
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.review.UIReview
@@ -9,8 +11,6 @@ import com.kirchhoff.movies.mapper.discover.IDiscoverMapper
 import com.kirchhoff.movies.mapper.review.IReviewListMapper
 import com.kirchhoff.movies.mapper.tv.ITvDetailsMapper
 import com.kirchhoff.movies.network.services.TvService
-import com.kirchhoff.movies.repository.BaseRepository
-import com.kirchhoff.movies.repository.Result
 
 class TvRepository(
     private val tvService: TvService,
