+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/PaginatedScreenVM.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.repository.Result
 import kotlinx.coroutines.launch
 
 abstract class PaginatedScreenVM<T : UIPaginated<*>> : ViewModel() {
