+++ app/src/test/java/com/kirchhoff/movies/mapper/discover/DiscoverMapperTest.kt
@@ -1,11 +1,11 @@
 package com.kirchhoff.movies.mapper.discover
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.data.ui.main.UIMovie
 import com.kirchhoff.movies.data.ui.main.UITv
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.main.NetworkMovie
 import com.kirchhoff.movies.networkdata.main.NetworkTv
-import com.kirchhoff.movies.repository.Result
 import com.kirchhoff.movies.utils.nextString
 import kotlin.random.Random
 import org.junit.Assert.assertEquals
