+++ app/src/test/java/com/kirchhoff/movies/mapper/person/main/PersonsMapperTest.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.mapper.person.main
 
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.main.NetworkPerson
-import com.kirchhoff.movies.repository.Result
 import com.kirchhoff.movies.utils.nextString
 import kotlin.random.Random
 import org.junit.Assert.assertEquals
