+++ app/src/test/java/com/kirchhoff/movies/ui/screens/details/movie/MovieDetailsVMTest.kt
@@ -1,10 +1,10 @@
 package com.kirchhoff.movies.ui.screens.details.movie
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.details.movie.UIMovieDetails
 import com.kirchhoff.movies.data.ui.details.movie.UITrailersList
-import com.kirchhoff.movies.repository.Result
 import com.kirchhoff.movies.repository.movie.IMovieRepository
 import com.kirchhoff.movies.utils.CoroutineRule
 import kotlinx.coroutines.ExperimentalCoroutinesApi
