+++ app/src/test/java/com/kirchhoff/movies/ui/screens/reviews/list/ReviewsListVMTest.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.ui.screens.reviews.list
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.review.UIReview
-import com.kirchhoff.movies.repository.Result
 import com.kirchhoff.movies.repository.movie.IMovieRepository
 import com.kirchhoff.movies.repository.tv.ITvRepository
 import com.kirchhoff.movies.ui.screens.core.verifyFirstPageErrorState
